+++ tests/rest/test_job.py
@@ -1,10 +1,11 @@
 import re
+from pathlib import Path
+
+import pytest
 
 import openeo
 from openeo.rest import JobFailedException, OpenEoClientException
-import pytest
-
-from openeo.rest.job import RESTJob
+from openeo.rest.job import RESTJob, ResultAsset
 from .. import as_path
 
 API_URL = "https://oeo.test"
@@ -133,7 +134,10 @@ def test_create_job_100(con100, requests_mock):
         }
         return True
 
-    requests_mock.post(API_URL + "/jobs", status_code=201, headers={"OpenEO-Identifier": "f00ba5"}, additional_matcher=check_request)
+    requests_mock.post(
+        API_URL + "/jobs",
+        status_code=201, headers={"OpenEO-Identifier": "f00ba5"}, additional_matcher=check_request
+    )
     con100.create_job({"foo1": {"process_id": "foo"}}, title="Foo", description="just testing")
 
 
@@ -165,6 +169,19 @@ def test_download_result(con100, requests_mock, tmp_path):
         assert f.read() == TIFF_CONTENT
 
 
+def test_get_results_download_file(con100, requests_mock, tmp_path):
+    requests_mock.get(API_URL + "/jobs/jj/results", json={"assets": {
+        "1.tiff": {"href": API_URL + "/dl/jjr1.tiff"},
+    }})
+    requests_mock.get(API_URL + "/dl/jjr1.tiff", content=TIFF_CONTENT)
+    job = RESTJob("jj", connection=con100)
+    target = as_path(tmp_path / "result.tiff")
+    res = job.get_results().download_file(target)
+    assert res == target
+    with target.open("rb") as f:
+        assert f.read() == TIFF_CONTENT
+
+
 def test_download_result_folder(con100, requests_mock, tmp_path):
     requests_mock.get(API_URL + "/jobs/jj/results", json={"assets": {
         "1.tiff": {"href": API_URL + "/dl/jjr1.tiff"},
@@ -181,6 +198,21 @@ def test_download_result_folder(con100, requests_mock, tmp_path):
         assert f.read() == TIFF_CONTENT
 
 
+def test_get_results_download_file_to_folder(con100, requests_mock, tmp_path):
+    requests_mock.get(API_URL + "/jobs/jj/results", json={"assets": {
+        "1.tiff": {"href": API_URL + "/dl/jjr1.tiff"},
+    }})
+    requests_mock.get(API_URL + "/dl/jjr1.tiff", content=TIFF_CONTENT)
+    job = RESTJob("jj", connection=con100)
+    target = as_path(tmp_path / "folder")
+    target.mkdir()
+    res = job.get_results().download_file(target)
+    assert res == target / "1.tiff"
+    assert list(p.name for p in target.iterdir()) == ["1.tiff"]
+    with (target / "1.tiff").open("rb") as f:
+        assert f.read() == TIFF_CONTENT
+
+
 def test_download_result_multiple(con100, requests_mock, tmp_path):
     requests_mock.get(API_URL + "/jobs/jj/results", json={"assets": {
         "1.tiff": {"href": API_URL + "/dl/jjr1.tiff"},
@@ -189,6 +221,32 @@ def test_download_result_multiple(con100, requests_mock, tmp_path):
     job = RESTJob("jj", connection=con100)
     with pytest.raises(OpenEoClientException, match="Expected one result file to download, but got 2"):
         job.download_result(tmp_path / "res.tiff")
+
+
+def test_get_results_multiple_download_single(con100, requests_mock, tmp_path):
+    requests_mock.get(API_URL + "/jobs/jj/results", json={"assets": {
+        "1.tiff": {"href": API_URL + "/dl/jjr1.tiff"},
+        "2.tiff": {"href": API_URL + "/dl/jjr2.tiff"},
+    }})
+    job = RESTJob("jj", connection=con100)
+    expected = r"Failed to get single asset \(name None\) from \['1.tiff', '2.tiff'\]"
+    with pytest.raises(OpenEoClientException, match=expected):
+        job.get_results().download_file(tmp_path / "res.tiff")
+
+
+def test_get_results_multiple_download_single_by_name(con100, requests_mock, tmp_path):
+    requests_mock.get(API_URL + "/jobs/jj/results", json={"assets": {
+        "1.tiff": {"href": API_URL + "/dl/jjr1.tiff"},
+        "2.tiff": {"href": API_URL + "/dl/jjr2.tiff"},
+    }})
+    requests_mock.get(API_URL + "/dl/jjr1.tiff", content=TIFF_CONTENT)
+    job = RESTJob("jj", connection=con100)
+    target = tmp_path / "res.tiff"
+    path = job.get_results().download_file(target, name="1.tiff")
+    assert path == target
+    assert list(p.name for p in tmp_path.iterdir()) == ["res.tiff"]
+    with (path).open("rb") as f:
+        assert f.read() == TIFF_CONTENT
 
 
 def test_download_results_040(session040, requests_mock, tmp_path):
@@ -235,3 +293,78 @@ def test_download_results(con100, requests_mock, tmp_path):
     assert set(p.name for p in target.iterdir()) == {"1.tiff", "2.tiff"}
     with (target / "1.tiff").open("rb") as f:
         assert f.read() == TIFF_CONTENT
+
+
+def test_get_results_download_files(con100, requests_mock, tmp_path):
+    requests_mock.get(API_URL + "/jobs/jj/results", json={"assets": {
+        "1.tiff": {"href": API_URL + "/dl/jjr1.tiff", "type": "image/tiff; application=geotiff"},
+        "2.tiff": {"href": API_URL + "/dl/jjr2.tiff", "type": "image/tiff; application=geotiff"},
+    }})
+    requests_mock.get(API_URL + "/dl/jjr1.tiff", content=TIFF_CONTENT)
+    requests_mock.get(API_URL + "/dl/jjr2.tiff", content=TIFF_CONTENT)
+    job = RESTJob("jj", connection=con100)
+    assert job.list_results() == {'assets': {
+        '1.tiff': {'href': 'https://oeo.test/dl/jjr1.tiff', 'type': 'image/tiff; application=geotiff'},
+        '2.tiff': {'href': 'https://oeo.test/dl/jjr2.tiff', 'type': 'image/tiff; application=geotiff'}
+    }}
+    target = as_path(tmp_path / "folder")
+    target.mkdir()
+    results = job.get_results()
+    downloads = results.download_files(target)
+    assert set(downloads) == {target / "1.tiff", target / "2.tiff"}
+    assert set(p.name for p in target.iterdir()) == {"1.tiff", "2.tiff"}
+    with (target / "1.tiff").open("rb") as f:
+        assert f.read() == TIFF_CONTENT
+
+
+def test_result_asset_download_file(con100, requests_mock, tmp_path):
+    href = API_URL + "/dl/jjr1.tiff"
+    requests_mock.get(href, content=TIFF_CONTENT)
+
+    job = RESTJob("jj", connection=con100)
+    asset = ResultAsset(job, name="1.tiff", href=href, metadata={'type': 'image/tiff; application=geotiff'})
+    target = as_path(tmp_path / "res.tiff")
+    path = asset.download(target)
+
+    assert isinstance(path, Path)
+    assert path.name == "res.tiff"
+    with path.open("rb") as f:
+        assert f.read() == TIFF_CONTENT
+
+
+def test_result_asset_download_folder(con100, requests_mock, tmp_path):
+    href = API_URL + "/dl/jjr1.tiff"
+    requests_mock.get(href, content=TIFF_CONTENT)
+
+    job = RESTJob("jj", connection=con100)
+    asset = ResultAsset(job, name="1.tiff", href=href, metadata={"type": "image/tiff; application=geotiff"})
+    target = as_path(tmp_path / "folder")
+    target.mkdir()
+    path = asset.download(target)
+
+    assert isinstance(path, Path)
+    assert path.name == "1.tiff"
+    with path.open("rb") as f:
+        assert f.read() == TIFF_CONTENT
+
+
+def test_result_asset_load_json(con100, requests_mock):
+    href = API_URL + "/dl/jjr1.json"
+    requests_mock.get(href, json={"bands": [1, 2, 3]})
+
+    job = RESTJob("jj", connection=con100)
+    asset = ResultAsset(job, name="out.json", href=href, metadata={"type": "application/json"})
+    res = asset.load_json()
+
+    assert res == {"bands": [1, 2, 3]}
+
+
+def test_result_asset_load_bytes(con100, requests_mock):
+    href = API_URL + "/dl/jjr1.tiff"
+    requests_mock.get(href, content=TIFF_CONTENT)
+
+    job = RESTJob("jj", connection=con100)
+    asset = ResultAsset(job, name="out.tiff", href=href, metadata={"type": "image/tiff; application=geotiff"})
+    res = asset.load_bytes()
+
+    assert res == TIFF_CONTENT